requirements.txt
# Flask Web Framework
Flask==2.3.3

# Flask Extensions
Flask-Mail==0.9.1

# Production Server
gunicorn==21.2.0

# Environment Management
python-dotenv==1.0.0

# Security
Werkzeug==2.3.7

# Additional utilities that might be useful
email-validator==2.0.0
WTForms==3.0.1
Flask-WTF==1.1.1

# Development dependencies (optional)
# Uncomment these for development
# flask-debugtoolbar==0.13.1
# pytest==7.4.2
# pytest-flask==1.2.0
